requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
python-dotenv==1.1.1
supabase==2.18.2
stripe==12.5.1
requests==2.32.4
